% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_numeric.R
\name{calculate_max_cycle_length}
\alias{calculate_max_cycle_length}
\title{Calculate the Maximum Possible Value of a Time Cycle}
\usage{
calculate_max_cycle_length(time_cycle_length, time_cycle_scale)
}
\arguments{
\item{time_cycle_length}{A character string indicating the larger unit of time.
It must be one of "year", "month", "day", "hour", "minute".}

\item{time_cycle_scale}{A character string indicating the smaller unit of time,
which is a division of the time_cycle_length. If time_cycle_length is "year",
time_cycle_scale can be one of "month", "day", "hour", "minute", "second".
If time_cycle_length is "month", time_cycle_scale can be "day", "hour", "minute",
"second", and so on.}
}
\value{
The function returns the maximum possible value that the time_cycle_scale
can take within one time_cycle_length.
}
\description{
This function returns the maximum possible value that a specific time
component can take, given a cycle length and scale.
}
\examples{

calculate_max_cycle_length("year", "month")  # returns 12
calculate_max_cycle_length("day", "minute")  # returns 1440
calculate_max_cycle_length("year", "day")  # returns 1440

}
