% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_numeric.R
\name{convert_date_to_numeric}
\alias{convert_date_to_numeric}
\title{Convert date to numeric value based on time cycle}
\usage{
convert_date_to_numeric(date_column, cycle_length, time_scale)
}
\arguments{
\item{date_column}{(Date or POSIXct) The column of dates to be converted.}

\item{cycle_length}{(character or integer) The length of the time cycle.
If character, it should be one of: 'year', 'month', 'day', 'hour', 'minute', 'second'.
If numeric, it represents the length of the cycle in arbitrary units.}

\item{time_scale}{(character or integer) The scale of the time cycle.
If character, it should be one of: 'year', 'month', 'day', 'hour', 'minute', 'second'.
If numeric, it represents the scale of the cycle in arbitrary units.}
}
\value{
A numeric vector with the same length as \code{date_column},
where each date has been converted to a numeric value based on the specified time cycle and scale.
}
\description{
This function converts a date or datetime column to a numeric format
based on a specified time cycle length and scale.
}
\examples{

convert_date_to_numeric(Sys.Date(), "year", "day") # Day of the year
convert_date_to_numeric(Sys.time(), "day", "hour") # Hour of the day

}
