% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw.R
\name{print.twdtw}
\alias{print.twdtw}
\title{Print method for twdtw class}
\usage{
\method{print}{twdtw}(x, ...)
}
\arguments{
\item{x}{An object of class \code{twdtw}}

\item{...}{Arguments passed to \code{\link{print.default}} or other methods.}
}
\description{
Print method for twdtw class
}
