% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_numeric.R
\name{to_date_time}
\alias{to_date_time}
\title{Convert to Date or POSIXct Format}
\usage{
to_date_time(x)
}
\arguments{
\item{x}{A vector of strings or objects to be converted.}
}
\value{
A vector of dates or datetimes.
}
\description{
This function converts a vector of date-like strings to Date or POSIXct format,
depending on the format of the input strings. It checks if the input is already
in Date or POSIXct format and performs the conversion accordingly.
}
\examples{

dates <- c("2023-07-15", "2023-07-16")

datetimes <- c("2023-07-15 12:30:00", "2023-07-16 13:45:00")

to_date_time(dates)

to_date_time(datetimes)

}
