% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdtw.R
\name{twdtw}
\alias{twdtw}
\title{Calculate TWDTW Distance between Time Series}
\usage{
twdtw(
  x,
  y,
  tw = c(100, 1),
  step_matrix = symmetric1,
  twdtw_version = "f77",
  index_column = "date",
  lower_band = TRUE,
  all_matches = FALSE
)
}
\arguments{
\item{x}{A tibble, data.frame, or data.table representing the first time series.}

\item{y}{A tibble, data.frame, or data.table representing the second time series.}

\item{tw}{Numeric vector of length 2 representing the time-weight parameters alpha and beta.
Defaults to c(100, 1) (set weight to zero).}

\item{step_matrix}{A matrix specifying the step pattern for the TWDTW algorithm.
Defaults to symmetric1.}

\item{twdtw_version}{A string identifying the version of twdtw implementation.
Options are 'f77' for Fortran 77, 'f90' for Fortran 90, 'cpp' for C++. Defaults to 'f77'.}

\item{index_column}{The column name of the time index. Defaults to "date".}

\item{lower_band}{Logical indicating whether to constrain the TWDTW calculation
to the lower band given by the time-weight parameter beta. Defaults to TRUE.}

\item{all_matches}{Logical indicating whether to find all matches within the
TWDTW matrix. Defaults to FALSE.}
}
\value{
A numeric value representing the TWDTW distance between the two time series.
}
\description{
This function calculates the Time-Weighted Dynamic Time Warping (TWDTW) distance
between two time series.
}
\examples{

n <- 20
t <- seq(0, pi, length.out = n)

x <- data.frame(date = seq(as.Date("2020-01-01"), by = "day", length.out = n),
             value = sin(t)*2 + runif(n))

y <- data.frame(date = seq(as.Date("2020-01-05"), by = "day", length.out = n),
             value = sin(t)*2 + runif(n))

rbenchmark::benchmark(
f90=twdtw(x, y, tw = c(-.1,50), all_matches = TRUE, twdtw_version = 'f90'),
fgt=twdtw(x, y, tw = c(-.1,50), all_matches = TRUE, twdtw_version = 'f90gt'),
cpp=twdtw(x, y, tw = c(-.1,50), all_matches = TRUE, twdtw_version = 'cpp'),
replications = 100
)

}
